Learned the Login and Register

+++ base/views.py
@@ -2,6 +2,63 @@ from django.shortcuts import render, redirect
 from django.db.models import Q
 from .models import Room,Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User 
+from django.contrib import messages
+from django.contrib.auth import authenticate , login , logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+
+def loginPage(request):
+    
+    page='login'
+    if( request.user.is_authenticated ):
+        return redirect('home')
+    
+    if( request.method == 'POST'):
+        username=request.POST.get('username').lower()
+        password=request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User Does Not Exist')
+        
+        user = authenticate(request, username=username, password = password)
+        
+        if user is not None:
+            login(request, user) #Create session in Database and Browser
+            return redirect('home')
+        else:
+            messages.error(request,'Credentials Wrong')
+            
+        
+    context={'page':page}
+    return render(request,'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    
+    #page='register'
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user= form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An Error Occured during registration")
+            
+    context = {'form':form}
+    return render(request,'base/login_register.html', context)
 
 def home(request):
     
@@ -21,6 +78,9 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+
+
+@login_required(login_url='login')
 def createRoom(request):
     
     form = RoomForm()
@@ -33,11 +93,15 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request,pk):
-    
     
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) #PreFilled the RoomForm with room Value
+    
+    if request.user!= room.host:
+        return HttpResponse('You are not Allowed Here!!')
     
     if(request.method=="POST"):
         form = RoomForm(request.POST,instance=room)
@@ -47,9 +111,15 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     
     room = Room.objects.get(id=pk)
+    
+    if request.user!= room.host:
+        return HttpResponse('You are not Allowed Here!!')
+    
     if request.method=='POST':
         room.delete()
         return redirect('home')
